move file + rename assert fn
