+++ apps/bot/src/domains/crew/utils/build-change-captain-view.ts
@@ -2,14 +2,12 @@ import type { Player } from '@one-piece/db';
 import { ActionRowBuilder, ButtonBuilder, ButtonStyle } from 'discord.js';
 import chunk from 'lodash/chunk.js';
 
-import { DISCORD_ACTION_ROW_MAX_BUTTONS, DISCORD_MESSAGE_MAX_BUTTONS } from '../../discord/constants.js';
-import type { View } from '../../discord/types.js';
-import { buildCustomId } from '../../discord/utils/build-custom-id.js';
-import { buildOpEmbed } from '../../discord/utils/build-op-embed.js';
-import type { CharacterRow } from '../character/types.js';
-import { getCharacterInstanceName } from '../character/utils/index.js';
-
-import { SET_CAPTAIN_BUTTON_NAME } from './constants.js';
+import { DISCORD_ACTION_ROW_MAX_BUTTONS, DISCORD_MESSAGE_MAX_BUTTONS } from '../../../discord/constants.js';
+import type { View } from '../../../discord/types.js';
+import { buildCustomId, buildOpEmbed } from '../../../discord/utils/index.js';
+import type { CharacterRow } from '../../character/types.js';
+import { getCharacterInstanceName } from '../../character/utils/index.js';
+import { SET_CAPTAIN_BUTTON_NAME } from '../constants.js';
 
 export function buildSetCaptainView(player: Player, crew: Array<CharacterRow>): View {
   const visibleCrew = crew.slice(0, DISCORD_MESSAGE_MAX_BUTTONS);
